requirements.txt
altair==5.4.0
attrs==24.2.0
av==10.0.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
entrypoints==0.4
filelock==3.16.0
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
imageio-ffmpeg==0.4.9
importlib-metadata==8.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numpy==1.26.4
openai-whisper==20231117
packaging==24.1
pandas==2.2.2
Pillow==10.4.0
protobuf==4.25.4
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pylibsrtp==0.10.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scipy==1.11.4
semver==3.0.2
six==1.16.0
smmap==5.0.1
sounddevice==0.4.7
streamlit==1.37.1
streamlit-webrtc==0.47.1
tenacity==9.0.0
toml==0.10.2
toolz==0.12.1
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.28.3
watchdog==4.0.1
zipp==3.20.0
